+++ dotfiles/.config/ml4w/scripts/installupdates.sh
@@ -45,8 +45,9 @@ sleep 1
 clear
 figlet -f smslant "Updates"
 echo
-
-if gum confirm "DO YOU WANT TO START THE UPDATE NOW?"; then
+primarycolor=$(cat ~/.config/ml4w/colors/primary)
+onsurfacecolor=$(cat ~/.config/ml4w/colors/onsurface)
+if gum confirm --selected.background=$primarycolor --prompt.foreground=$onsurfacecolor "DO YOU WANT TO START THE UPDATE NOW?"; then
     echo
     echo ":: Update started..."
 elif [ $? -eq 130 ]; then
@@ -63,8 +64,22 @@ fi
 
 # Arch
 if [[ $(_checkCommandExists "pacman") == 0 ]]; then
-    aur_helper="$(cat ~/.config/ml4w/settings/aur.sh)"
-    $aur_helper
+
+    yay_installed="false"
+    paru_installed="false"
+    if [[ $(_checkCommandExists "yay") == 0 ]]; then
+        yay_installed="true"
+    fi
+    if [[ $(_checkCommandExists "paru") == 0 ]]; then
+        paru_installed="true"
+    fi
+    if [[ $yay_installed == "true" ]] && [[ $paru_installed == "false" ]]; then
+        yay
+    elif [[ $yay_installed == "false" ]] && [[ $paru_installed == "true" ]]; then
+        paru -Syu --noconfirm
+    else
+        yay
+    fi
 
 # Fedora
 elif [[ $(_checkCommandExists "dnf") == 0 ]]; then