+++ setup/_fonts.sh
@@ -5,3 +5,5 @@
 
 sudo cp -rf $SCRIPT_DIR/fonts/FiraCode /usr/share/fonts
 sudo cp -rf $SCRIPT_DIR/fonts/Fira_Sans /usr/share/fonts
+sudo cp -rf $SCRIPT_DIR/fonts/BebasNeue-Regular /usr/share/fonts
+sudo cp -rf $SCRIPT_DIR/fonts/Material-Icons /usr/share/fonts